moved MarkdownTextArea to common components
