+++ anki/src/features/editmode/components/Description.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { ButtonSwitch } from "components/ButtonSwitch";
-import { MarkdownTextArea } from "./MarkdownTextArea";
+import { MarkdownTextArea } from "components/MarkdownTextArea";
 import { DeckStuffDTO } from "../types";
 
 type Props = {
